Add CSP, authenticated encryption, drop serverkey and some other bits
Content-Security-Policy needs some work -- probably need to add some
JavaScript build process, as currently it's manually updated in
app.psgi.

Dropping serverkey by default because this is just an extended salt,
using PBKDF2 now means it isn't really needed. (We achieve this with
minimal changes by keeping the empty line on the .txt endpoint.)

Authentication is achieved by HMAC-SHA512 over the ciphertext, with a
key derived from the secretkey, in the same way the AES key is derived
from the key with a salt mixed in. While HKDF would be better this keeps
compatibility with the OpenSSL scheme. (Implementing authentication in
the CLI is tricky because "openssl mac" needs the key material passing
on the command line, which would make things more insecure.)